Run dom0 tests with virtual screen (xvfb)
Tests window was loosing focus due to an unkown reason (possibly a test
opening a ghost window). This was causing issues where the remainder of
the test would fail to run due to xterm no longer being in focus, thus
impossible to type in. For context, see https://github.com/freedomofpress/securedrop-workstation/issues/1411